Reorganisation based on rubygems guide
