+++ lib/rspec/teamcity/formatter_initializer.rb
