+++ lib/rspec/teamcity/rake_exceptions.rb
