+++ lib/rspec/teamcity/rakerunner_consts.rb
