+++ lib/rspec/teamcity/runner_common.rb
