+++ lib/rspec/teamcity/utils/logger_util.rb
