+++ lib/rspec/teamcity/utils/runner_utils.rb
