+++ lib/rspec/teamcity/utils/service_message_factory.rb
