+++ lib/rspec/teamcity/utils/std_capture_helper.rb
