+++ lib/rspec/teamcity/utils/string_ext.rb
