+++ lib/rspec/teamcity/utils/url_formatter.rb
