+++ lib/rspec/teamcity_rspec3.rb
